Deduplicate shared syn helpers so the Rust clone gate stays clean.
Extract pattern_name, string_literal, and read_rust_syntax for the rules
merged onto main after rebase, keeping the zero-debt duplicate-code gate.
